scale bug is removed
